removed quotes from variable, shellcheck complains
The shellcheck program to check the syntax of this script wants double quotes around the ${s_ytdl_options} variable, but the youtube-dl program just will not be having it. shellcheck is afraid of globbing in the variable, but youtube-dl thinks the variable text is a single option. Quotes removed. I can deal with a single non-error from shellcheck.
*facepalm*